backend/telegram-bot/requirements.txt
openai>=1.0.0,<2.0.0
psycopg2-binary
requests
openpyxl>=3.1.0